agents/ingestion-agent/requirements.txt
# Cloud Function Runtime
functions-framework>=3.0.0

# Dependencies from Shared Library (Must match source/shared/requirements.txt)
grpcio>=1.60.0
google-cloud-firestore>=2.13.0
google-cloud-secret-manager>=2.16.0
google-cloud-storage>=2.14.0
google-genai>=0.3.0
google-cloud-pubsub>=2.18.0
requests>=2.31.0
aiohttp>=3.9.0
cryptography>=41.0.0
dataclasses-json>=0.6.0
pydantic>=2.9.0
typing-extensions>=4.12.0
